[refactor] entity table 컬렉션 일급컬렉션으로 변경
